% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_containers.R
\name{create_container_from_repo}
\alias{create_container_from_repo}
\title{Create container based on repository}
\usage{
create_container_from_repo(contribution_row)
}
\arguments{
\item{contribution_row}{}
}
\description{
Create container based on repository
}
