% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_containers.R
\name{create_containers}
\alias{create_containers}
\title{Create Docker containers for database}
\usage{
create_containers(all_contributions)
}
\arguments{
\item{all_contributions}{database}
}
\description{
Create Docker containers for database
}
