% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_contributions.R
\name{donwload_single_contribution}
\alias{donwload_single_contribution}
\title{Download single contribution}
\usage{
donwload_single_contribution(contribution_row)
}
\arguments{
\item{contribution_row}{database row}
}
\description{
Download single contribution
}
