% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_contributions.R
\name{download_contributions}
\alias{download_contributions}
\title{Download all contributions from database}
\usage{
download_contributions(all_contributions)
}
\arguments{
\item{all_contributions}{database}
}
\description{
Download all contributions from database
}
