% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_info_to_contributions.R
\name{get_last_commit_sha}
\alias{get_last_commit_sha}
\title{Get last Git commit id}
\usage{
get_last_commit_sha(contribution_row)
}
\arguments{
\item{contribution_row}{database row}
}
\description{
Get last Git commit id
}
