% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_info_to_contributions.R
\name{git_info_to_contributions}
\alias{git_info_to_contributions}
\title{Add Git information to database}
\usage{
git_info_to_contributions(all_contributions)
}
\arguments{
\item{all_contributions}{database}
}
\description{
Add Git information to database
}
