% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Main function to get third party resources}
\usage{
main(
  content_contributions_filename = "content-contributions.csv",
  all_cards_filename = "zettelkasten.csv",
  source_dir = "."
)
}
\arguments{
\item{content_contributions_filename}{Database with contributors}
}
\description{
Main function to get third party resources
}
