% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_contributions.R
\name{prepare_contributions}
\alias{prepare_contributions}
\title{Prepare contributions database}
\usage{
prepare_contributions(all_contributions)
}
\arguments{
\item{all_contributions}{database}
}
\value{
database
}
\description{
Prepare contributions database
}
