% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_contributions.R
\name{render_contributions}
\alias{render_contributions}
\title{Render all contributions from database}
\usage{
render_contributions(all_contributions)
}
\arguments{
\item{all_contributions}{}
}
\description{
Render all contributions from database
}
