% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_contributions.R
\name{render_single_contribution}
\alias{render_single_contribution}
\title{Render single contribution}
\usage{
render_single_contribution(contribution_row)
}
\arguments{
\item{contribution_row}{}
}
\description{
Render single contribution
}
